+++ src/core/mod.rs
@@ -3,3 +3,4 @@ pub mod metrics;
 pub mod neighbor;
 pub mod node;
 pub mod parameters;
+pub mod calc;
